Added support for ofxOnsetDetection (used in showSegmentClcGirls1 right now).
Laid down groundwork for support for multiple segments, with multiple effects in single show. ShowDefinition gets a list of showSegments, and a pointer to the currentSegment. The segment will become responsible for the actual drawing of the frame, because there are so many details involved that cannot be abstracted away (clearing the depth buffer between passes, setting blend states, etc). This means that the draw and render functions will be taken away from MeshView.

The implementation of the standard visualization that comes with mapamok, with the example shader, will be moved to the showSegmentDefault class.

Each showSegment will get its own control panel. Which control panel is visible will depend on the active segment. If it wants, a segment can create multiple control panels, for example, one for each effect that is present in the segment.